requirements.txt
requests
PyYAML
argparse